+++ changelog.md
@@ -1,6 +1,7 @@
 ##### Version 1.20.4-11.5.0-pre
 - Add warning signs (BluSunrize)
     - Textures kindly provided by AriaElidove, thank you!
+- Add fence gates for treated wood, steel and aluminium (BluSunrize)
 - Remove raw ore blocks from the list of blocks dug by the grinding disk  (voidsong-dragonfly)
 - Change workbench to be able to handle blueprints with more than 9 recipes (BluSunrize)
 - Change villager trades for revolver pieces and mineral deposit maps to re-roll (BluSunrize)